LC-valid-perfect-square [Runtime: N/A; Memory: N/A]

+++ binary-search/valid-perfect-square.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    bool isPerfectSquare(int num) {
-      int n = num/2;
+    bool isPerfectSquare(long num) {
+      long n = num/2;
       for(int i=0 ; i<=num ; i++)
       {
         if(i*i == num) return true;
